keiba_dual/requirements.txt
pandas
numpy
scikit-learn
xgboost
requests
python-dotenv
joblib
